+++ components/src/LibraryBook/Elements/AdultBadge.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const AdultBadge: React.SFC = () => (
+  <p>19세 미만 구독불가</p>
+);
