+++ components/src/LibraryBook/Elements/Author.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const Author: React.SFC<{ author: string }> = (props) => (
+  <p>{ props.author }</p>
+);
