+++ components/src/LibraryBook/Elements/DownloadSize.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const DownloadSize: React.SFC<{ downloadSize: number }> = (props) => (
+  <p>{ props.downloadSize }MB</p>
+);
