+++ components/src/LibraryBook/Elements/Expired.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const Expired: React.SFC = () => (
+  <p>기간 만료</p>
+);
