+++ components/src/LibraryBook/Elements/ExpiredAt.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+export interface ExpiredAtProps {
+  expiredAt: string;
+}
+
+export const ExpiredAt: React.SFC<ExpiredAtProps> = (props) => (
+  <p>{ props.expiredAt }</p>
+);
