+++ components/src/LibraryBook/Elements/NotAvailable.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const NotAvailable: React.SFC = () => (
+  <p>이용불가</p>
+);
