+++ components/src/LibraryBook/Elements/ReadingProgressBar.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+export interface ReadingStatusProps {
+  readingProgress?: number;
+}
+
+export const ReadingProgressBar: React.SFC<ReadingStatusProps> = (props) => (
+  <p>독서진행률 { props.readingProgress }%</p>
+);
