+++ components/src/LibraryBook/Elements/Ridiselect.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const Ridiselect: React.SFC = () => (
+  <p>리디셀렉트</p>
+);
