+++ components/src/LibraryBook/Elements/ThumbnailImage.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+
+export interface ThumbnailImageProps {
+  thumbnailUrl: string;
+  thumbnailSize?: number;
+  title: string;
+}
+
+export const ThumbnailImage: React.SFC<ThumbnailImageProps> = (props) => (
+  <img src={props.thumbnailUrl} alt={props.title} />
+);
