+++ components/src/LibraryBook/Elements/Title.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const Title: React.SFC<{ title: string }> = (props) => (
+  <p>{ props.title }</p>
+);
