+++ components/src/LibraryBook/Elements/UnReadDot.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const UnReadDot: React.SFC = () => (
+  <p>아직 읽지 않은 책</p>
+);
