+++ components/src/LibraryBook/Elements/UnitBookDownloading.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const UnitBookDownloading: React.SFC = () => (
+  <p>유닛도서 다운로드 중</p>
+);
