+++ components/src/LibraryBook/Elements/UpdateBadge.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const UpdateBadge: React.SFC = () => (
+  <p>신규 업데이트</p>
+);
