+++ components/src/LibraryBook/Thumbnail/index.tsx
@@ -1,44 +1,23 @@
 import * as React from 'react';
-
 import {
-  // enum
-  AdditionalInfoProps,
-  AdultBadge,
-  BadgeProps,
-  // interface
-  BookCountProps,
-  Checkbox,
   DOWNLOAD_STATUS,
-  DownloadButton,
-  DownloadStatusProps,
-  Expired,
-  ExpiredAt,
-  NotAvailable,
-  // components
   READING_STATUS,
-  ReadingProgressBar,
-  ReadingStatusProps,
-  Ridiselect,
-  SelectProps,
-  ThumbnailImage,
-  ThumbnailImageProps,
-  UnitBookDownloadingProgressCircle,
-  UnitLinkButton,
-  UnReadDot,
-  UpdateBadge,
   VIEW_TYPE,
-  ViewTypeProp,
-} from './components';
+} from '../constants';
+import * as LibraryBook from '../Elements';
 
 export interface ThumbnailProps extends
-  ViewTypeProp,
-  ThumbnailImageProps,
-  BadgeProps,
-  SelectProps,
-  BookCountProps,
-  AdditionalInfoProps,
-  DownloadStatusProps,
-  ReadingStatusProps {
+  LibraryBook.ThumbnailImageProps,
+  LibraryBook.SelectProps,
+  LibraryBook.BookCountProps,
+  LibraryBook.DownloadStatusProps,
+  LibraryBook.ReadingStatusProps {
+    available?: boolean;
+    unitBook?: boolean;
+    expiredAt?: string;
+    ridiSelect?: boolean;
+    adultBadge?: boolean;
+    updateBadge?: boolean;
     bookId: string;
     children?: React.ReactNode;
     [extraKey: string]: any;
@@ -72,45 +51,45 @@ export const Thumbnail: React.SFC<ThumbnailProps> = (props) => {
     <div
       { ...extraProps }
     >
-      { editMode && <Checkbox /> }
+      { editMode && <LibraryBook.Checkbox /> }
       { readingStatus && <>
-        { readingStatus === READING_STATUS.New && <UnReadDot /> }
+        { readingStatus === READING_STATUS.New && <LibraryBook.UnReadDot /> }
         { readingStatus === READING_STATUS.Opened && viewType === VIEW_TYPE.Portrait &&
-          <ReadingProgressBar readingProgress={readingProgress} />
+          <LibraryBook.ReadingProgressBar readingProgress={readingProgress} />
         }
       </> }
-      <ThumbnailImage
+      <LibraryBook.ThumbnailImage
         thumbnailUrl={ thumbnailUrl }
         title={ title }
       />
-      { adultBadge && <AdultBadge /> }
-      { updateBadge && <UpdateBadge /> }
+      { adultBadge && <LibraryBook.AdultBadge /> }
+      { updateBadge && <LibraryBook.UpdateBadge /> }
       { viewType === VIEW_TYPE.Portrait &&
         <div>
           { unitBook && <>
-            <UnitLinkButton
+            <LibraryBook.UnitLinkButton
               bookCount={bookCount}
               bookCountUnit={bookCountUnit}
             />
             { downloadStatus === DOWNLOAD_STATUS.Downloading &&
-              <UnitBookDownloadingProgressCircle/>
+              <LibraryBook.UnitBookDownloading/>
             }
           </> }
           { !unitBook && <>
             { available &&
-              <DownloadButton
+              <LibraryBook.DownloadButton
                 downloadStatus={downloadStatus}
                 downloadProgress={downloadProgress}
               />
             }
-            { ridiSelect && <Ridiselect /> }
-            { !ridiSelect && !available && <Expired /> }
-            { !ridiSelect && available && expiredAt && <ExpiredAt expiredAt={ expiredAt } /> }
+            { ridiSelect && <LibraryBook.Ridiselect /> }
+            { !ridiSelect && !available && <LibraryBook.Expired /> }
+            { !ridiSelect && available && expiredAt && <LibraryBook.ExpiredAt expiredAt={ expiredAt } /> }
           </> }
         </div>
       }
       { children }
-      { !available && <NotAvailable /> }
+      { !available && <LibraryBook.NotAvailable /> }
     </div>
   );
 };
